add basic optional resource discovery
Adds a very basic discovery of optional resource creation/enablement
toggles in the values.yaml JSONSchema for the Helm Chart. We look
through the properties in the JSONSchema for anything that is "create"
or "enabled" and guess that this values.yaml configuration path is a
toggle (in a Helm template) to create a particular Kubernetes resource.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>