auto adjust KubeVersion for Helm Chart rendering
The Helm Chart may specify a KubeVersion in its metadata that is
incompatible with the Kubernetes client version used in compiling the
Helm Go SDK. If this is the case, we need to pass an updated KubeVersion
installer option when rendering.

This patch adds code to automatically determine a good KubeVersion if
the KubeVersion of the Helm SDK is incompatible with the version
constraint in the Helm Chart's metadata.KubeVersion value.

Closes Issue #2

Signed-off-by: Jay Pipes <jaypipes@gmail.com>